Move the code for the first pass to its own module
